pyproject.toml
[tool.poetry]
name = "adk-hackathon"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohappyeyeballs = "2.6.1"
aiohttp = "3.11.18"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
anyio = "4.9.0"
appnope = "0.1.4"
argon2-cffi = "25.1.0"
argon2-cffi-bindings = "21.2.0"
arrow = "1.3.0"
asttokens = "3.0.0"
async-lru = "2.0.5"
attrs = "25.3.0"
authlib = "1.6.0"
babel = "2.17.0"
beautifulsoup4 = "4.13.4"
bleach = "6.2.0"
cachetools = "5.5.2"
certifi = "2025.4.26"
cffi = "1.17.1"
charset-normalizer = "3.4.2"
click = "8.2.1"
comm = "0.2.2"
cryptography = "45.0.2"
debugpy = "1.8.14"
decorator = "5.2.1"
defusedxml = "0.7.1"
deprecated = "1.2.18"
distro = "1.9.0"
docstring-parser = "0.16"
executing = "2.2.0"
faker = "37.3.0"
fastapi = "0.115.12"
fastjsonschema = "2.21.1"
fqdn = "1.5.1"
frozenlist = "1.6.0"
google-adk = "1.0.0"
google-api-core = "2.24.2"
google-api-python-client = "2.170.0"
google-auth = "2.40.2"
google-auth-httplib2 = "0.2.0"
google-cloud-aiplatform = "1.94.0"
google-cloud-bigquery = "3.33.0"
google-cloud-core = "2.4.3"
google-cloud-resource-manager = "1.14.2"
google-cloud-secret-manager = "2.23.3"
google-cloud-speech = "2.32.0"
google-cloud-storage = "2.19.0"
google-cloud-trace = "1.16.1"
google-crc32c = "1.7.1"
google-genai = "1.16.1"
google-resumable-media = "2.7.2"
googleapis-common-protos = "1.70.0"
graphviz = "0.20.3"
grpc-google-iam-v1 = "0.14.2"
grpcio = "1.71.0"
grpcio-status = "1.71.0"
h11 = "0.16.0"
httpcore = "1.0.9"
httplib2 = "0.22.0"
httpx = "0.28.1"
httpx-sse = "0.4.0"
idna = "3.10"
importlib-metadata = "8.6.1"
ipykernel = "6.29.5"
ipython = "9.3.0"
ipython-pygments-lexers = "1.1.1"
isoduration = "20.11.0"
jedi = "0.19.2"
jinja2 = "3.1.6"
jiter = "0.10.0"
json5 = "0.12.0"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
jsonschema = "4.24.0"
jsonschema-specifications = "2025.4.1"
jupyter-events = "0.12.0"
jupyter-lsp = "2.2.5"
jupyter-client = "8.6.3"
jupyter-core = "5.8.1"
jupyter-server = "2.16.0"
jupyter-server-terminals = "0.5.3"
jupyterlab = "4.4.3"
jupyterlab-pygments = "0.3.0"
jupyterlab-server = "2.27.3"
langchain = "0.3.25"
langchain-core = "0.3.61"
langchain-text-splitters = "0.3.8"
langsmith = "0.3.42"
markupsafe = "3.0.2"
matplotlib-inline = "0.1.7"
mcp = "1.9.1"
mistune = "3.1.3"
multidict = "6.4.4"
nbclient = "0.10.2"
nbconvert = "7.16.6"
nbformat = "5.10.4"
nest-asyncio = "1.6.0"
notebook = "7.4.3"
notebook-shim = "0.2.4"
numpy = "2.2.6"
openai = "1.82.0"
opentelemetry-api = "1.33.1"
opentelemetry-exporter-gcp-trace = "1.9.0"
opentelemetry-resourcedetector-gcp = "1.9.0a0"
opentelemetry-sdk = "1.33.1"
opentelemetry-semantic-conventions = "0.54b1"
orjson = "3.10.18"
overrides = "7.7.0"
packaging = "24.2"
pandas = "2.3.0"
pandocfilters = "1.5.1"
parso = "0.8.4"
pexpect = "4.9.0"
platformdirs = "4.3.8"
prometheus-client = "0.22.1"
prompt-toolkit = "3.0.51"
propcache = "0.3.1"
proto-plus = "1.26.1"
protobuf = "5.29.4"
psutil = "7.0.0"
ptyprocess = "0.7.0"
pure-eval = "0.2.3"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.2"
pycparser = "2.22"
pydantic = "2.11.5"
pydantic-settings = "2.9.1"
pydantic-core = "2.33.2"
pygments = "2.19.1"
pyparsing = "3.2.3"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.1.0"
python-json-logger = "3.3.0"
python-multipart = "0.0.20"
pytz = "2025.2"
pyyaml = "6.0.2"
pyzmq = "26.4.0"
referencing = "0.36.2"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rfc3339-validator = "0.1.4"
rfc3986-validator = "0.1.1"
rpds-py = "0.25.1"
rsa = "4.9.1"
send2trash = "1.8.3"
setuptools = "80.9.0"
shapely = "2.1.1"
six = "1.17.0"
sniffio = "1.3.1"
soupsieve = "2.7"
sqlalchemy = "2.0.41"
sse-starlette = "2.3.5"
stack-data = "0.6.3"
starlette = "0.46.2"
tenacity = "9.1.2"
terminado = "0.18.1"
tinycss2 = "1.4.0"
tornado = "6.5.1"
tqdm = "4.67.1"
traitlets = "5.14.3"
types-python-dateutil = "2.9.0.20250516"
typing-inspection = "0.4.1"
typing-extensions = "4.13.2"
tzdata = "2025.2"
tzlocal = "5.3.1"
uri-template = "1.3.0"
uritemplate = "4.1.1"
urllib3 = "2.4.0"
uvicorn = "0.34.2"
wcwidth = "0.2.13"
webcolors = "24.11.1"
webencodings = "0.5.1"
websocket-client = "1.8.0"
websockets = "15.0.1"
wrapt = "1.17.2"
yarl = "1.20.0"
zipp = "3.21.0"
zstandard = "0.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
